% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cereal.R
\docType{data}
\name{Cereal}
\alias{Cereal}
\title{Cereal data}
\format{The variables are
\itemize{
\item \code{name}: The brand name of the cereal.
\item \code{calories}: Number of calories per serving.
\item \code{protein}: Number of grams of protein per serving.
\item \code{fat}: Number of grams of fat per serving.
\item \code{sodium}: Milligrams of sodium per serving.
\item \code{fiber}: Number of grams of dietary fiber per serving.
\item \code{carbo}: Number of grams of complex carbohydrates per serving.
\item \code{sugars}: Number of grams of sugar per serving.
\item \code{potass}: Milligrams of potassium per serving.
\item \code{shelf}: Display shelf counting from the floor.  A higher number is a higher shelf.
}}
\source{
Each company's website.
}
\usage{
data(Cereal)
}
\description{
Cereal data
}
\details{
This data measures different nutritional values of all
all the brands of cold cereal along with the shelf the cereal
occupies in the store.
}
\keyword{datasets}
