% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChicagoTaxi.R
\docType{data}
\name{ChicagoTaxi}
\alias{ChicagoTaxi}
\title{Chicago taxi data}
\format{The variables are
\itemize{
\item \code{Trip}: The total number of seconds the passenger is in the taxi.
\item \code{Fare}: The total fare (measured in dollars)
\item \code{Payment}: The payment method (cash or credit card)
}}
\source{
data.cityofchicago.org
}
\usage{
data(ChicagoTaxi)
}
\description{
Chicago taxi data
}
\details{
This data measures the total trip time (in seconds), the cost of the ride (in dollars), and the payment method.
}
\keyword{datasets}
