% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollegeData.R
\docType{data}
\name{CollegeData}
\alias{CollegeData}
\title{Payscale Return on Investment in College Data}
\format{The variables are
\itemize{
\item \code{Rank}: Rank of the school in terms of ROI.
\item \code{Name}: Name of the school.
\item \code{ROI}: Twenty year net return on investement.  It is the difference between 20 year median pay for a bachelor's grad and 24 year median pay for a high school grad minus total four-year cost.
\item \code{Cost}: Full cost, including tuition and fees, room & board, and books&supplies.
\item \code{GradRate}: Percent of full-time, first-time degree seeking undergraduate students who began their studies at the given school and graduate within six years.
\item \code{Years}: Number of years it takes for at least 65 percent of graduates from the school to receive their diplomas.
\item \code{Loan}: Average loan amount for full-time,first-time degree-seeking undgraduate students.
}}
\source{
\url{payscale.com/college-roi#roi-list-bottom-content}
}
\usage{
data(CollegeData)
}
\description{
Payscale Return on Investment in College Data
}
\details{
This is data provided by \code{payscale.com} that
measures various variables that may be used to judge the
monetary return on an investment in a college education.
}
\keyword{datasets}
