% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diamonds.R
\docType{data}
\name{Diamonds}
\alias{Diamonds}
\title{Diamond data}
\format{The variables are
\itemize{
\item \code{Carat}: The weight of the diamond.
\item \code{Color}: The color of the diamond is graded. The lower the letter, the more colorless the diamond.
\item \code{Clarity}: The clarity is a measure of the number of blemishes in a diamond.  FL: Flawless IF: Internally Flawless WS1: Very, Very Slightly Included 1  WS2: Very, Very Slightly Included 2 VS1: Very Slightly Included 1  VS2: Very Slight Included 2 SI1: Slightly Included 1  SI2: Slightly Included 2 I1: Inclusions 1  I2: Inclusions 2
\item \code{Cut}: A measure of the quality of the diamond's angles, brilliance, and finishing details.
\item \code{Price}: The price of the diamond (in dollars).
}}
\source{
diamonds.com
}
\usage{
data(Diamonds)
}
\description{
Diamond data
}
\details{
This data measures the carat, color, clarity, cut, and price of a random sample of diamonds.
}
\keyword{datasets}
