% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomeRun_2017.R
\docType{data}
\name{HomeRun_2017}
\alias{HomeRun_2017}
\title{HomeRun_2017 data}
\format{The variables are
\itemize{
\item \code{Date}: Date the home run was hit.
\item \code{Hitter}: The name of the player who hit the home run.
\item \code{HitterTeam}: The team of the player who hit the home run.
\item \code{Pitcher}: The name of the pitcher who gave up the home run.
\item \code{PitcherTeam}: The team of the pitcher who gave up the home run.
\item \code{Inning}: The inning the home run was hit.
\item \code{Ballpark}: The name of the baseball park where the home run was hit.
\item \code{Distance}: Distance the home run traveled (measured in feet)
\item \code{SpeedOffBat}: Speed of the ball, in miles per hour, when the ball left the bat.
\item \code{ElevAngle}: Launch angle of the ball.
\item \code{HorizAngle}: The initial direction of the ball as it left the bat in degrees, where 45 degrees is straight down the right field line, 90 degrees is straight over second base and 135 degrees is straight down the left field line.
\item \code{Type}: The type of home run.  JE: Just enough - home run barely cleared the fence by less than 10 vertical feet, OR that it landed less than one fence height past the fence.  ND: No doubt- which means the ball cleared the fence by at least 20 vertical feet AND landed at least 50 feet past the fence. ITP: Inside the park home run. Ball did not travel over the outfield fence. PL: Plenty (all the rest)
}}
\source{
baseballsavant.mlb.com
}
\usage{
data(HomeRun_2017)
}
\description{
HomeRun_2017 data
}
\details{
This data measures variables for all home runs hit in the 2017 Major League baseball season.
}
\keyword{datasets}
