% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SullivanStatsSurveyI.R
\docType{data}
\name{SullivanStatsSurveyI}
\alias{SullivanStatsSurveyI}
\title{Sullivan Statistics Survey I data}
\format{The variables are
\itemize{
\item \code{Gender}: The gender of the respondent.
\item \code{Age}: The age of the respondent (in years).
\item \code{Birth}: The birth month of the respondent.
\item \code{Text}: In the past week, have you texted while driving?
\item \code{Language}: Is a language other than English the primary language spoken in your home?
\item \code{TV}: How many televisions are in your home?
\item \code{Smoke}: Do you currently smoke?
\item \code{Party}: What best describes your political affiliation?
\item \code{Philosophy}: What best describes your political philosophy?
\item \code{Hour_Internet}: How many hours did you spend on the Internet last week (Sunday through Saturday)?
\item \code{Hour_TV}: How many hours did you watch television last week (Sunday through Saturday)?
\item \code{Rich}: In your opinion, whta minimum annual income for an individual qualifies as an individual to be considered rich?
\item \code{Deficit}: Would you be willing to pay higher taxes if the tax revenue went directly toward deficit reduction?
\item \code{Tickets}: How many speeding tickets have you been issued in the past 12 months?
\item \code{Children}: What do you think is the ideal number of children for a family to have?
\item \code{Height}: What is your height (in inches)?
\item \code{Foot}: What is the length of your foot (in inches)?
\item \code{Job}: Do you currently have a job?
\item \code{Hour_Work}: If yes, how many hours did you work last week (Sunday through Saturday)?
\item \code{Credit}: Do you currently have any credit cards?
\item \code{Cards}: If yes, how many credit cards do you have?
\item \code{Debt}: if yes, to the nearest dollar, what is the total outstanding balance on your credit cards?

@keywords datasets

@source StatCrunch Survey
}}
\usage{
data(SullivanStatsSurveyI)
}
\description{
Sullivan Statistics Survey I data
}
\details{
This data represents the results of a survey written by the students of Michael Sullivan.
}
\keyword{datasets}
