% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SullivanStatsSurveyII.R
\docType{data}
\name{SullivanStatsSurveyII}
\alias{SullivanStatsSurveyII}
\title{Sullivan Statistics Survey II data}
\format{The variables are
\itemize{
\item \code{Gender}: The gender of the respondent.
\item \code{Age}: The age of the respondent (in years).
\item \code{Education}: What is your level of education?
\item \code{Tax}: What percent of income do you believe individuals should pay in federal income tax?
\item \code{Inequality}: Do you believe there an income inequality discrepancy between males and females when each has the same experience and education?
\item \code{MinWage}: Do you beleive there should be a minimum wage?
\item \code{Amount}: If yes what do you believe the minimum wage should be?
\item \code{Philosophy}: What is your political philosophy?
\item \code{Text}: Do you believe it is okay to text while at a red light (as the driver)?
\item \code{RetirementDollars}: How much do you believe (in today's dollars) you will need for a comfortable retirement?
\item \code{Retire}: What do you believe is the ideal retirement age?
\item \code{Death}: For how many years would you like to live?

@keywords datasets

@source StatCrunch Survey
}}
\usage{
data(SullivanStatsSurveyII)
}
\description{
Sullivan Statistics Survey II data
}
\details{
This data represents the results of a survey written by the students of Michael Sullivan.
}
\keyword{datasets}
