% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tornadoes_2017.R
\docType{data}
\name{Tornadoes_2017}
\alias{Tornadoes_2017}
\title{Tornado data}
\format{The variables are
\itemize{
\item \code{Year}: The year of the tornado.
\item \code{Month}:  The month of the tornado.
\item \code{State}: State in which the tornado occurred.
\item \code{F_Scale}: F-scale rating on the Fujita scale.
\item \code{Injuries}: Number of injuries as a result of the tornado.
\item \code{Fatalities}: Number of fatalities as a result of the tornado.
\item \code{Length}: The number of miles the tornado is on the ground.
\item \code{Width}: The width of the tornado in yards.
\item \code{Number_States}: The number of states in which the tornado was on the ground.

@keywords datasets

@source spc.noaa.gov/wcm/#data
}}
\usage{
data(Tornadoes_2017)
}
\description{
Tornado data
}
\details{
This data represents a variety of variables measured on all tornadoes that struck the United States between 1950 and 2018.
}
\keyword{datasets}
